+++ src/cernml/lsa_utils/__init__.py
@@ -39,6 +39,13 @@ up.
 """
 
 try:
+    from ._hooks import (
+        AbstractHooks,
+        DefaultHooks,
+        Hooks,
+        InconsistentHookInstalls,
+        get_current_hooks,
+    )
     from ._incorporator import Incorporator, IncorporatorGroup, NotFound
     from ._utils import (
         get_context_by_user,
@@ -51,10 +58,15 @@ except ImportError as exc:  # pragma: no cover
     raise ImportError("import this package in `with LSAClient().java_api()`") from exc
 
 __all__ = [
+    "AbstractHooks",
+    "DefaultHooks",
+    "Hooks",
+    "InconsistentHookInstalls",
     "Incorporator",
     "IncorporatorGroup",
     "NotFound",
     "get_context_by_user",
+    "get_current_hooks",
     "get_cycle_type_attributes",
     "get_settings_function",
     "incorporate_and_trim",